refactor: Move ChromeRuntimeRewriteRuleRepository to persistence directory
- Move repository implementation from messaging/ to persistence/
- Update import paths in contentContainer.ts
- Move and update test files accordingly

ChromeRuntimeRewriteRuleRepository is a repository implementation that
uses messaging, not a messaging service itself. Persistence directory
is the appropriate location for repository implementations.
