+++ host-frontend-root/frontend-src-root/tests/unit/infrastructure/di/contentContainer/concrete-class-registration-completeness.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 import { ApplyRulesOnDomMutationUseCase } from 'src/application/usecases/contentOnMessageReceived/ApplyRulesOnDomMutationUseCase';
 import { GetElementSelectionUseCase } from 'src/application/usecases/selection/GetElementSelectionUseCase';
 import { contentContainer } from 'src/frameworks-and-drivers/di/contentContainer';
-import { ChromeRuntimeRewriteRuleRepository } from 'src/frameworks-and-drivers/messaging/ChromeRuntimeRewriteRuleRepository';
+import { ChromeRuntimeRewriteRuleRepository } from 'src/frameworks-and-drivers/persistence/ChromeRuntimeRewriteRuleRepository';
 import { DebounceTimer } from 'src/infrastructure/browser/timer/DebounceTimer';
 import { WindowCurrentUrlService } from 'src/infrastructure/browser/window/WindowCurrentUrlService';
 import { DomRootChecker } from 'src/infrastructure/document/DomRootChecker';
